--- UMM/Extension/View+Extension.swift
@@ -1,8 +1,0 @@
-//
-//  View+Extension.swift
-//  UMM
-//
-//  Created by Wonil Lee on 10/22/23.
-//
-
-import SwiftUI
